Refactor, extract utility functions
