+++ src/index.js
@@ -1,12 +1,15 @@
 const stylelint = require('stylelint')
+const { validBase, hasPixelValue, validPixelValue } = require('./utils')
+
 const { createPlugin } = stylelint
 const { ruleMessages, validateOptions, report } = stylelint.utils
 
+const ruleName = 'plugin/8-point-grid'
 const rules = {
-  'plugin/8-point-grid': {
-    base: 8,
-  },
+  [ruleName]: { base: 8 },
 }
+
+const plugins = ['stylelint-8-point-grid']
 
 const blacklist = [
   'margin',
@@ -35,50 +38,21 @@ const blacklist = [
   'left',
 ]
 
-const plugins = ['stylelint-8-point-grid']
+// ignore values with `calc` and sass variables
+const unsupported = ['calc', '\\$\\w+']
 
-const ruleName = 'plugin/8-point-grid'
+const pattern = (props) =>
+  new RegExp(props.map((prop) => `^${prop}$`).join('|'))
+
+const unsupportedPattern = new RegExp(unsupported.join('|'))
+
+const valid = (val) =>
+  hasPixelValue(val) && !String(val).match(unsupportedPattern)
 
 const messages = ruleMessages(ruleName, {
   invalid: (prop, actual, base) =>
     `Invalid \`${prop}: ${actual}\`. It should be divisible by ${base}.`,
 })
-
-const pattern = (props) =>
-  new RegExp(props.map((prop) => '^' + prop + '$').join('|'))
-
-const validBase = (val) => !isNaN(parseFloat(val)) && isFinite(val)
-
-const hasPxValue = (val) => String(val).includes('px')
-
-const validPixelValue = (value, base, whitelist) => {
-  return (
-    // handle multiple px values
-    //   e.g. padding: 8px 8px 1px 8px or padding: 3em 8px 8px 8px;
-    value
-      .split(/[\s\r\n]+/)
-      .filter(hasPxValue)
-      .every(
-        (value) => isWhitelist(whitelist, value) || divisibleBy(value, base)
-      )
-  )
-}
-
-// ignore values with `calc` and sass variables
-const unsupported = ['calc', '\\$\\w+']
-
-const unsupportedPattern = new RegExp(unsupported.join('|'))
-
-const valid = (val) => hasPxValue(val) && !String(val).match(unsupportedPattern)
-
-const isWhitelist = (whitelist, value) =>
-  (whitelist && whitelist.includes(value)) || false
-
-const divisibleBy = (value, base) => {
-  const number = value.match(/\d+/)
-  if (isNaN(number)) return true
-  return Number(number) % Number(base) === 0
-}
 
 const rule = createPlugin(ruleName, (primaryOption) => {
   return (postcssRoot, postcssResult) => {
@@ -87,7 +61,7 @@ const rule = createPlugin(ruleName, (primaryOption) => {
       possible: {
         base: validBase,
         ignore: blacklist,
-        whitelist: hasPxValue,
+        whitelist: hasPixelValue,
       },
     })
     if (!validOptions) return
